New gen/ directory with massive overhaul: generating optimized C code
Judicious use of C and -O2 allows us to generate most of the code for
the primitives in a portable and efficient way. Then most of the
wrapping code can be written in C.

The end result should be (at least) as performant as the original
hand-edited asm version in portable/ and vastly more flexible for
changes (eg. moving TOS into a register) thanks to the codegen.